pyproject.toml
[project]
name = "drf-board"
version = "0.1.0"
description = ""
authors = [
    {name = "SWWS97",email = "parkmon2@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-filter (>=25.1,<26.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
